refactor(models, scripts): models配下のフォルダ構成整理・type統合とcopy-schemaスクリプト修正
