+++ packages/models/src/frontend/index.ts
@@ -1,5 +1,6 @@
 import * as v from "valibot";
 import { ISODateSchema } from "../common";
+export * from "./type";
 
 export const JobOverviewSchema = v.object({
   jobNumber: v.string(),
@@ -19,3 +20,4 @@ export const JobDetailSchema = v.object({
   workingHours: v.string(),
   qualifications: v.string(),
 });
+
