+++ packages/models/src/frontend/type.ts
@@ -1,8 +1,5 @@
+import type { JobDetailSchema, JobOverviewSchema } from ".";
 import type * as v from "valibot";
-import type {
-  JobDetailSchema,
-  JobOverviewSchema,
-} from "../../schemas/frontend";
-
 export type TJobOverview = v.InferOutput<typeof JobOverviewSchema>;
 export type TJobDetail = v.InferOutput<typeof JobDetailSchema>;
+
