+++ packages/models/src/headless-crawler/index.ts
@@ -1,1 +1,2 @@
 export * from "./scraper";
+export * from "./type";
