+++ packages/models/src/headless-crawler/scraper.ts
@@ -1,3 +1,4 @@
+import type { Locator, Page } from "playwright";
 import * as v from "valibot";
 export const jobNumberSchema = v.pipe(
   v.string(),
@@ -178,3 +179,4 @@ export const ScrapedJobSchema = v.object({
   jobDescription: jobDescriptionSchema,
   qualifications: v.nullable(qualificationsSchema),
 });
+
