+++ packages/models/src/job-store/index.ts
@@ -3,3 +3,4 @@ export * from "./jobFetch";
 export * from "./jobInsert";
 export * from "./jobList";
 export * from "./jobList/continue";
+export * from "./type"
