--- packages/models/src/types/headless-crawler/crawler-type.ts
@@ -1,7 +1,0 @@
-import type { JobNumber } from "./shared-type";
-
-export type JobMetadata = {
-  jobNumber: JobNumber;
-};
-
-export type NewJobOpeningsFilter = "TodayYesterday" | "Within1Week";
