--- packages/models/src/types/headless-crawler/index.ts
@@ -1,3 +1,0 @@
-export * from "./crawler-type";
-export * from "./scraper-type";
-export * from "./shared-type";
