--- packages/models/src/types/headless-crawler/scraper-type.ts
@@ -1,8 +1,0 @@
-import type { Page } from "playwright";
-import type * as v from "valibot";
-import type { ScrapedJobSchema } from "../../schemas";
-
-const jobDetailPage = Symbol();
-export type JobDetailPage = Page & { [jobDetailPage]: unknown };
-
-export type ScrapedJob = v.InferOutput<typeof ScrapedJobSchema>;
