--- packages/models/src/types/index.ts
@@ -1,3 +1,0 @@
-export * from "./frontend";
-export * from "./headless-crawler";
-export * from "./job-store";
