--- packages/models/src/types/job-store/jobInsert.ts
@@ -1,6 +1,0 @@
-import type { InferOutput } from "valibot";
-import type { insertJobRequestBodySchema } from "../../schemas";
-
-export type InsertJobRequestBody = InferOutput<
-  typeof insertJobRequestBodySchema
->;
